Fixed issue where a ck editor field wouldn't update xml value on change and on switching to backend.
Fixed issue with img search and help tool where returning a single file would cause an issue. Fixed wikimedia file name encoding issue that was preventing some images from being displayed.

Adjusted wikimedia commons get request to keep trying until the desired amount of images is reached, and added filtering to stop non-image types like long pdfs making it into the img search and help tool results.

Enabled the html button for 'keeping' images so that it doesn't permanently grey out when a request fails or if the user exits out of the lightbox early.

+++ editor/imagesearchandhelp/imgSelection.php
@@ -2,19 +2,21 @@
 require_once (str_replace('\\', '/', __DIR__) . "/../../config.php");
 if(isset($_SESSION["paths_img_search"])){
 	$images = $_SESSION["paths_img_search"];
-	for($i = 0; $i < count($_POST["indices_to_delete"]); $i++){
-		$index = $_POST["indices_to_delete"][$i];
-		if($index > count($images) && $index < 0) {
-			continue;
-		}
-		$image = $images[$index];
-		$ext = pathinfo($image, PATHINFO_EXTENSION);
-		$credits = str_replace($ext, "txt", $image);
-		//echo $image . "<br>";
-		//x_check_path_traversal($image);
-		unlink($image);
-		unlink($credits);
-	}
+    if ($_POST["indices_to_delete"]!== null){
+        for($i = 0; $i < count($_POST["indices_to_delete"]); $i++){
+            $index = $_POST["indices_to_delete"][$i];
+            if($index > count($images) && $index < 0) {
+                continue;
+            }
+            $image = $images[$index];
+            $ext = pathinfo($image, PATHINFO_EXTENSION);
+            $credits = str_replace($ext, "txt", $image);
+            //echo $image . "<br>";
+            //x_check_path_traversal($image);
+            unlink($image);
+            unlink($credits);
+        }
+    }
 	unset($_SESSION["paths_img_search"]);
 	echo "OK";
 }
